backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
sqlalchemy>=2.0.0
pyjwt>=2.8.0
bcrypt>=4.1.0
python-multipart>=0.0.6
aiofiles>=23.2.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
websockets>=12.0
